get_item -> get_item_index + get_canvas
Make the former function get_item into a method named get_item_index.
The $c parameter was never used, so it is dropped, also the $y
parameter is optional and may be conveniently calculated to the
current event position.

Also a new method get_canvas was added to get to the "real" canvas.